move schema to schema.qraphl file
